Add match_case_calculator.py

+++ control-flow/match_case_calculator.py
@@ -2,8 +2,8 @@
 # A simple match case calculator
 
 # Take first and second numbers from user
-num1= int(input(" Enter the first number: "))
-num2 = int(input(" Enter the second number: "))
+num1 = float(input(" Enter the first number: "))
+num2 = float(input(" Enter the second number: "))
 
 # Ask user for preferred operation
 operation = input(" Choose the operation (+, -, *, /): " )
